+++ CHANGELOG.md
@@ -24,6 +24,11 @@ _If you are upgrading: please see [`UPGRADING.md`](UPGRADING.md#unreleased)._
 - 💥 Require generated device projects to use Python 3.11 or newer ([#515])
   ([\@denialhaag])
 
+### Fixed
+
+- 🐛 Preserve timestamps of unchanged prefixed headers to avoid redundant builds
+  ([#537]) ([\@burgholzer])
+
 ## [1.3.3] - 2026-08-19
 
 _If you are upgrading: please see [`UPGRADING.md`](UPGRADING.md#133)._
@@ -231,6 +236,7 @@ for previous changelogs._
 <!-- PR links -->
 
 [#538]: https://github.com/Munich-Quantum-Software-Stack/QDMI/pull/538
+[#537]: https://github.com/Munich-Quantum-Software-Stack/QDMI/pull/537
 [#515]: https://github.com/Munich-Quantum-Software-Stack/QDMI/pull/515
 [#486]: https://github.com/Munich-Quantum-Software-Stack/QDMI/pull/486
 [#485]: https://github.com/Munich-Quantum-Software-Stack/QDMI/pull/485